src/Digital_human_command_interface/pyproject.toml

[project]
name = "daishan-new"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.104.0",
    "httpx>=0.25.0",
    "openai>=2.15.0",
    "psutil>=5.9.0",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0",
    "ragflow-sdk>=0.0.1",
    "requests>=2.31.0",
    "uvicorn[standard]>=0.24.0",
]
